pyproject.toml
[project]
name = "gcp-mcp"
version = "0.1.0"
description = "Experimental MCP server for interacting with Google Cloud Platform resources"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.0",
    "google-cloud-secret-manager>=2.23.3",
    "google-cloud-run>=0.10.18",
    "google-cloud-resource-manager>=1.14.2",
    "google-cloud-logging>=3.8.0",
    "pytest>=8.3.5",
    "fastmcp>=2.4.0",
]

[project.scripts]
gcp-mcp = "gcp:mcp.run"

[tool.ruff]
target-version = "py313"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["."]
sources = ["."]
